+++ NEWS.md
@@ -1,7 +1,8 @@
 # *News*
 
 # zipcodeR 0.1.1.9000 (WIP)
-Add function for obtaining congressional districts of a given ZIP code.
+- Add function for obtaining congressional districts of a given ZIP code.
+- Add function for obtaining ZIP codes contained within a congressional district.
 
 # zipcodeR 0.1.0 (2020-09-08)
 Initial public release, first version accepted by CRAN.